refact: move UI 'ui' package
