+++ feature/gallery/src/androidTest/java/com/jeanbarrossilva/orca/feature/gallery/ui/test/zoom/SemanticsNodeInteraction.extensions.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.feature.gallery.test.zoom
+package com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom
 
 import androidx.compose.ui.semantics.SemanticsNode
 import androidx.compose.ui.test.SemanticsNodeInteraction
