+++ feature/gallery/src/androidTest/java/com/jeanbarrossilva/orca/feature/gallery/ui/test/zoom/TouchInjectionScope.extensions.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.feature.gallery.test.zoom
+package com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom
 
 import androidx.compose.ui.test.TouchInjectionScope
 import androidx.compose.ui.test.pinch
