+++ feature/gallery/src/androidTest/java/com/jeanbarrossilva/orca/feature/gallery/ui/zoom/TouchInjectionScopeExtensionsTests.kt
@@ -13,18 +13,18 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.feature.gallery.zoom
+package com.jeanbarrossilva.orca.feature.gallery.ui.zoom
 
 import androidx.compose.ui.test.TouchInjectionScope
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.performTouchInput
-import com.jeanbarrossilva.orca.feature.gallery.Gallery
-import com.jeanbarrossilva.orca.feature.gallery.test.onPage
-import com.jeanbarrossilva.orca.feature.gallery.test.zoom.assertIsZoomedIn
-import com.jeanbarrossilva.orca.feature.gallery.test.zoom.assertIsZoomedOut
-import com.jeanbarrossilva.orca.feature.gallery.test.zoom.performZoomIn
-import com.jeanbarrossilva.orca.feature.gallery.test.zoom.zoomIn
-import com.jeanbarrossilva.orca.feature.gallery.test.zoom.zoomOut
+import com.jeanbarrossilva.orca.feature.gallery.ui.Gallery
+import com.jeanbarrossilva.orca.feature.gallery.ui.test.onPage
+import com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom.assertIsZoomedIn
+import com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom.assertIsZoomedOut
+import com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom.performZoomIn
+import com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom.zoomIn
+import com.jeanbarrossilva.orca.feature.gallery.ui.test.zoom.zoomOut
 import com.jeanbarrossilva.orca.platform.autos.theme.AutosTheme
 import org.junit.Rule
 import org.junit.Test
